requirements.txt
adafruit-io
pillow
pyinaturalist
